Implemation first Demo version of mtq_controller Including: informers for the new CRDs,Migrations,VMIs and for Kubernetes components. Logic for status update of the new CRD. cmd/mtq_controller.go as a starting point for the new controller.
Description of new VirtualMachineMigrationResourceQuota CRD:
- VirtualMachineMigrationResourceQuota is namespaced
- Spec:
AdditionalMigrationResources:Define the maximum resources allowed
to be exceeded during migrations.

- Status:
MigrationsToBlockingResourceQuotas:Map that include:
[Migration] -> {set of resourceQuota in the namespace
that are blocking us and can be unblocked by adding
resource of AdditionalMigrationResources}

OriginalBlockingResourceQuotas: include set of
Modifed ResourceQuotas with the original spec
These resourceQuotas should be restored once
The target pod is created

The controller is triggered when a migration object
is created\deleted\updated.

*In the future the controller will be trigger when
VirtualMachineMigrationResourceQuota is created\updated
with sync loop that includes all the migrations in the namespace.

if the migration object is blocked than the controller increase
and can be unblocked by increasing resource quota with
AdditionalMigrationResources ,it will increase the resourceQuota
resources and restore it once the target virt-laucher pod is
created.

** in the future the namespace will be locked untill the
target pod is created so other pods won't be created
at the same time.
Signed-off-by: bmordeha <bmodeha@redhat.com>