Enhance VMMRQ with desired state in spec and actual state in status
This commit adds support for a desired state in the VMMRQ
specification (vmmrq.spec) and an actual state in the
VMMRQ status (vmmrq.status).
This enhancement enables users to modify the desired
state even when the namespace is locked, allowing
for more flexible management of resource allocation
during virtual machine migrations.

Signed-off-by: bmordeha <bmodeha@redhat.com>